requirements.txt
Django>=5.0,<6.0
psycopg2-binary>=2.9
dj-database-url>=2.1
redis>=5.0
celery>=5.3
django-celery-beat>=2.5
django-htmx>=1.17
django-tailwind>=3.8.0
django-browser-reload>=1.12
django-widget-tweaks>=1.5
python-decouple>=3.8
Pillow>=10.2
python-docx>=1.1
openai>=1.10
whitenoise>=6.6
gunicorn>=21.2
django-extensions>=3.2
ipython>=8.20
pytest-django>=4.7
coverage>=7.4
